Move all languages found by geoip to the top in Welcome spoke
There might be more languages used in a territory detected by GeoIP
and while anaconda preselects the most probable one, it has been
noted during the usability testing that even the less probable
languages should be also highlighted.

Therefore, all languages from geoip we have translations for
are now moved to the top of the language list in Welcome spoke,
above the separator. The most likely language is still on the top
of the list and is preselected.

Also for the given territory preselect the correct locale variant.
EXAMPLE: For the GB territory, the language should be preset to English
and the Locale to English(UK).