Translate layout and switching options descriptions on the fly
Creating a new instance of the XklWrapper is quite an expensive action as it
reads and processes a lot of data and creates a lot of internal structure. So
instead of creating new instance the get fresh translations on language change,
it's better to translate strings on the fly if they are requested.