Create the XklWrapper singleton in background
Creating the XklWrapper instance takes quite a lot of time (~4 seconds on a VM).
Instead of doing when the first spoke needs it, we can create the instance in
background as soon as the X server is started. Since it includes a lot of I/O
operations, threading makes a real difference (at least 2 seconds on a VM). But
the main advantage is that this long action happens before we show anything to
user and not when some grey screen is shown.