Mirror the GUI if an RTL language is chosen (#1008397)
Some languages (or more precisely scripts) use left-to-right reading direction,
some use right-to-left reading direction. The UI should reflect that fact and
widgets should be arranged as if being mirrored.

Gtk itself handles the right direction for a chosen language, but it is set only
when the toolkit is initialized (Gtk.init()). Since we change language in the
GUI we need to set the direction manually later.

Until Gtk exports a function to get the default direction for the chosen
language [1], we need to use the same hack as the Gtk itself uses internally.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1008821