Make progress screen in text mode standalone spoke instead of hub
The progress screen in text mode is not a hub. It cannot provide any spokes
because it prints a lot of text to the output and spokes would be rolled off in
a moment. It should be a standalone spoke which also fixes the issue with it not
being scheduled because of no spokes to be provided.

Resolves: rhbz#1008788