Fix the password confirmation match check (#1009907)
The check for whether the password field and the password confirmation
field match is run when either field is changed, since changes to either
field will cause the state of the match to change. This was working
correctly when a change to either field broke the match, but was not
always working when a change fixed the match. The check object for the
field not changed remained in a failure state, making it impossible to
exit the spoke even though the password matched. This fixes that by
manually running the check on the field not changed.