Add translation support to check_accelerators
Added a --translate option to check_accelerators.py. When set, it will
check all of the translated keyboard accelerators in addition to the
English ones. Changed the exception check so that language information
can be included as <!-- check_accelerators(<lang>): comment -->. Added
support for the new parameter to the run_check_accelerators.sh wrapper
script.

`make check' will not check translations.

Requires python-polib to parse the .po files.