Make Software spoke ready even if there is no repo (#1010348)
Nothing can be done in the Software spoke if the repository is not set up, but
automated installations wait for all spokes to become ready which never happens
if there is no repo specified in the kickstart. The Software spoke thus needs to
inform user what to do, but has to be ready/visitable.