Clear bootDisk and bootloader stage info on errors (#1013482)
If there is an error picking a bootloader stage1 there are a number of
variables that need to be cleared so that stale information isn't used.
This adds a new .reset() method to the bootloader that resets stage1 and
stage2 related items. And calls it before raising stage1 related
BootLoaderErrors.

The ksdata bootloader's bootDisk also need to be reset. It should be
setup after any autopart or custom changes, not before. Now that
un-partitioned disks can be used there are cases where a disk's stage1
status may change. eg. If a raw formatted disk may be partitioned, it
can then be used for stage1.

Also catch any kickstart errors that may show up because of changes
(eg. deselecting the disk used in an --onpart)

Resolves: rhbz#1013482