Don't try to collect removed modules
If some module is removed, the makeupdates scripts replaces it with a stub
raising an error to crash when testing. However, methods collecting spokes and
hubs go over such modules without explicitly referring them in the code (what
would be a bug). We should ignore such errors and just ignore those modules.