Only encrypt the TUI user password once (#1015220)
Since the TUI user spoke apply() method is called for every property
change, and cryptPassword was called in apply(), the user password was
being set multiple times if a password was set followed by other
properties being set (e.g., set password, then set administrator).

Moved the password encryption into EditTUIDialog so that cryptPassword
is called immediately after the prompts and an encrypted password is
returned to the spoke.