Exit on exception in the askVNC spoke (#962804)
Install a handler for exception messages from the TUI in askvnc. A
handler would normally be installed by TextUserInterface.setup(), but
askvnc runs before that.  There's not a lot we can do about exceptions
at this point, so display it and exit.

The previous failure mode was to loop forever attempting to reprint the
spoke.