Use different colors for different message types.
This is how the GtkInfoBar we have used to work, but at some point the
gnome-themes developers decided that they didn't like that anymore. See
https://bugzilla.gnome.org/show_bug.cgi?id=704266. The white-on-blue bar
is kind of terrible to read, and anaconda uses the warning and error
colors to attract attention to warnings and errors, so this style change
undoes the color changes and makes the infobar colors jarring again.