Add checks for unexpanded macros.
Autoconf checks for unexpanded macros, which usually indicates a missing
definition for that macro, but it only works if the macro names start
with AC/AS/AM and probably a couple of other things. The
gobject-introspection and gtk-doc macros start with
GOBJECT_INTROSPECTION and GTK_DOC, respectively, so added checks for
those two macros.