Only prompt for LUKS password if the user has chosen to configure automatically.
Resolves: #966510

Note: This change is the same as for bug #1017330, but in Fedora.

If the user has chosen any other action then they will be prompted for a
LUKS password after they complete that action.

In the case where they RECLAIM space on the disk they won't be prompted for
a password before they return to installation summary,
but they won't be done either, so they will have to return to storage
to complete.

If they cancel or quit there is no point in prompting them for a password,
because they have made no changes.

Signed-off-by: mulhern <amulhern@redhat.com>