Add storage tests.
These kind of fit into a weird area - they're blivet, but they're also
kickstart.  Thus, this seems like the best place for them.  You have to be
root to run them but if you're not root, "make check" will just skip over
them and not fail the test process.

The idea here is small, focused tests that verify anaconda behaves in a certain
way.  This includes checking that anaconda fails like it should in some cases,
and succeeds in others.  For now, we cannot verify that what was made matches
what we wanted.  That may come later, if needed.