Fix the alignment of the Network Time switch (#1019301)
The default for valign and halign is "fill", which matches the rest of
the widgets in the top row of controls, which is what we want. Remove
the alignment overrides which align the NTP switch to the top of the
header box instead of the middle.

(cherry picked from commit 73fa1de2a71d3a2221174fefaf0ba0509a51a865)