Use cache for base repo if possible
The baseRepo property is tricky, because it needs yum_lock to be evaluated.
However, in many cases, nothing changes in the configuration and we can simply
return the same value as in the previous call. If there is some change that can
potentially change the value, the cache needs to be made invalid and refreshed.

This speeds things up and makes the getBaseRepo method cheap enough to be used
in additional places that need it.