Replace placeholders with the strings from python
For strings that had keyboard accelerators defined in python but not in
glade, replace the glade placeholder with a non-translatable version of
the python string.

Split the labels on buttons and radio buttons into explicit children in
order to set the necessary markup attributes on the label. This also
cleans up some of that get_children mess we have with builder objects.