Fall back to closest mirror in source (#1016801)
method.method is special. None means closest mirror, and if we set it to
"url" it causes the metadata to always be re-fetched when exiting the
spoke.

This depends on a pykickstart fix for method getattr which isn't
returning url by default.

(cherry picked from commit e18e913981f1fcd68bbd7b58baf864a910074029)