Add support for context-based translations
This change adds three new functions to pyanaconda.i18n:

  C_ takes a context and a message and returns a translated message.
  This is analagous to _.

  CN_ takes a context and a message, marks them as translatable, and
  returns the original message. The message must be translated elsewhere
  using C_ and the context string that was passsed to CN_. This is
  analagous to N_.

  CP_ takes a context, a message, the plural form of the message, and a
  number and returns a translated message according to the pluralization
  rules of the target language. This is analagous to P_.

The reason for this is that strings in a user interface often involve
abbreviations that are sensitive to the other strings used within a
context. For example, consider the GUI string "_Add". The underscore
indicates that the 'A' can be used as a keyboard shortcut. This might be
translated to Spanish as "_Agregar". However, one of the windows where
"_Add" is used may also contain a stock "_OK" button, which is
translated to Spanish as "_Aceptar". By using a separate message context
for different instances of "_Add", the translator is able to set
different translations for different contexts and thus choose different
keyboard shortcuts where appropriate.