Don't update hub's continue button and label for every spoke (#1020373)
If a spoke is left, we need to update completeness of the spokes and also of the
hub. But instead of updating completeness of the hub after updating completeness
of every single spoke, we can do it afterwards only once. This removes a
possibly long hang (on live installations) caused by poking in yum and storage
many times.