Revert "Only prompt for LUKS password if the user has chosen to configure automatically."
Related to fedora #966510.

This change caused the opposite problem; sometimes the user is not prompted
for a LUKS password even when they ought to be. This causes a failure on the
Progress hub.

In any case, it seems that the place to prompt a user for the password should
possibly be somewhere else than on_back_clicked in StorageSpoke or CustomSpoke.
Ideally it should be later...on a spoke right after the user has pressed Begin
Install. That way, instead of taking into consideration where the user is going
and what screens they might possibly hit, the code need only account for the
choice the user has ultimately made: to encrypt or not.

This reverts commit 79561f847360bd91e155111ad57a30c50ee09c6f.