Avoid configure-event loops. (#1021511)
Only resize and move the lightbox if the geometry changed since the last
configure-event. Do not call gdk_window_restack: the restack triggers
another configure-event, these windows aren't supposed to be restackable
in the first place, and if the window manager does allow them to be
restacked there's not a whole lot that we can promise anyway.