No longer use summary screen visit to decide whether bootloader has been configured (#1025802)
Resolves: f20#1025802

Previously, the code incorporated the assumption that if there had been
a kickstart install, and no bootloader information had been specified in the
kickstart file, then there was insufficient information to configure the
bootloader. Therefore, the bootloader was not considered configured until the
user had examined the bootloader choices on the summary screen. Now the same
default location is supplied for either kickstart or manual install, so the
bootloader is always considered to have been configured, even if the user
has neither examined the summary screen or explicitly specified in the
kickstart file.

The line that sets the variable self.data.bootloader.seen = True is being left
in, because it records the state, bootloader choice has been seen, correctly.

(cherry-picked from commit c75e11e52deb61ea343b580741229da3869f732a)

Signed-off-by: mulhern <amulhern@redhat.com>