Correctly accept 'sshd' boot arg as alias for 'inst.sshd' (#924157)
In dracut/parse-anaconda-options.sh we have:

  check_depr_arg "sshd" "inst.sshd"

which makes us accept "sshd" as an alias for inst.sshd, *inside* the
initrd. But this script runs in the anaconda runtime, so we don't have
/etc/cmdline.d/75-anaconda-options.conf anymore, so getarg only sees
the actual boot args - which don't contain inst.sshd.

We could copy /run/install/cmdline.d/* into /etc/cmdline.d, but I'm not
sure if that would have surprising or unintended side-effects.

So let's do the simple thing: check for both inst.sshd and sshd.