Fix the layout up and down button sensitivies.
Disable both the up and down buttons when there is only one item in the
list. Update the top/bottom/one based sensitivity test when the list
changes.

While I was at it, store the layout selection object in the keyboard
spoke object so that we don't have to call get_object() in every signal
handler.