Add check for Linux HFS+ ESP on Mac (#1010495)
On Mac it can be difficult to tell the difference between the hfs+
partition that fedora creates to use as the EFI System Partition and a
normal OSX hfs+ partition.

This switches to using a special filesystem, named macefi, that is hfs+
formatted and has a partition label of 'Linux HFS+ ESP'.