No longer install anaconda user documentation (#1026448)
Resolves: fed#1026448

The current anaconda install puts some user documentation files in
/usr/share/doc/anaconda-<version>. The documents are now omitted from
the installation, because anaconda isn't your typical user program.
It seems more appropriate
to make that documentation available in the official docs and not
dependent on the install for users who are in the process of installing.

The ultimate goal is to make all user documents available externally
in the most official documentation that it is sensible to put it and use
the docs directory for developer documents.

Previously command-line.txt was
downloaded from the wiki as part of the build process,
but the page has moved so instead of a useful
document you get a 301 html page. Now, nothing is downloaded.

boot-options.html can be generated from boot-options.txt using asciidoc.
This is not made part of the build process because the ultimate goal is
to no longer maintain boot-options.txt and to have all that information
incorporated in the install guide instead.

Since command-line.txt will not be downloaded any more, it is removed from
cleans and ignores as appropriate.

install-methods.txt has been obsoleted and is removed.

(cherry-picked from commit 62ecdc74545dccaa7086c721f6f259d9489399b5)

Signed-off-by: mulhern <amulhern@redhat.com>