Fix Gtk errors about list store columns
The renderer for keyboard layouts was set to pull the foreground color
from column six of the store, but there's only one column, so Gtk was
spitting out "Gtk-CRITICAL: column < priv->n_columns" or something to
that effect.

The GtkCellRendererText editor in Glade is just a giant list of numbers,
so this was probably added by some misaimed scroll wheeling.