Set bootloader default location to mbr in constructor (#1025811)
Resolves: fed#1025811

Previously bootloader location defaulted to approximately none for a
kickstart install if there was no bootloader line in the file.
If there was a bootloader line in the file, it defaulted to "mbr".
For an interactive install, location was explicitly set
to "mbr" by the initial kickstart file.

Now, the default is "mbr" regardless of how the install is started.

(cherry-picked from commit 7114bd9308ea8684341d2b1a354d5e5050ae332d)

Signed-off-by: mulhern <amulhern@redhat.com>