Set environment variables in anaconda systemd shell file (#1023913)
Resolves: rhbz#1023913

Previously, if the user entered anaconda shell, HOME was not set. Now
it is set to /, as it was in RHEL6.

PATH is explicitly set to the value it was before this change as it turns
out that a side-effect of setting HOME to / is that PATH is initialized
differently. A directory, /usr/local/sbin,  is prepended to this PATH
as a side-effect of login. The directory does not exist, so this should
have no effect.

PS1 is explicitly set to its previous value for the same reason PATH is set.

Signed-off-by: mulhern <amulhern@redhat.com>