Make dialog return code checking more robust
A dialog can return not only the values associated with its buttons but also
some other values. The value that it can certainly return is -4. This value
is most likely to occur when the user presses the escape key, and causes
the dialog to be deleted.

In some cases this possibility is not allowed for and the wrong action is
taken when the escape key is pressed because the code puts the unexpected
value in an equivalence class with some button that actually causes an
action to be taken.

The *.py files in pyanaconda were searched for occurances of the regular
expression =.*run and nearby lines were scrutinized and corrected if
necessary.

In general, the changes make an if condition more restrictive if the then
branch causes an action and less restrictive if it does not.

The changes to summary.glade from a previous commit
are reversed and custom.py
is updated appropriately as using the ResponseType constants would
be inconsistant with the rest of the code.

Related: rhbz#998384

Signed-off-by: mulhern <amulhern@redhat.com>

Conflicts:
	pyanaconda/ui/gui/spokes/custom.py