Avoid raising an exception if epdb module unavailable (#1026779)
Resolves: rhbz#1026779

Catch the exception and print a log message if the import fails.
The startDebugging method is a convenience for developers, and this seemed
to be the most useful approach for those I talked to.

With this approach, if the developer wants
to load epdb into their updates.img file they can still use the HUP signal
functionality without further edits.

Signed-off-by: mulhern <amulhern@redhat.com>