Remove threading from getBaseRepo handling (#1011555)
The original purpose of the getBaseRepo cache was to ensure that
requests it weren't constantly sitting on the _yum_lock. This was
accomplished with a combination of a cache, decorators to invalidate the
cache and a thread to re-validate the cache.

The problem is that the thread can be fired off at any time, and when
this happens in the middle of a PAYLOAD_MD thread update it can change
the state of yum's preconf/conf attributes, resulting in a traceback
the next time it tries to access .preconf

This patchset hopefully simplifies things a bit by removing the
threading from the picture. Now what happens is:

Any method decorated with @invalidates_br_cache will re-validate the
cache when the method is *finished* and if it meets the criteria
passed to the decorator.

If getBaseRepo is called and the cache is invalid it will immediately
re-validate it -- all this does is examine repos and their enabled
attribute, this should not be a long running process.