Add function to map functions on items in the main thread
Sometimes we need to map a function on a bunch of items in way that it is called
in the main thread. And sometimes we also need to do some transformations on the
items before we do the actions that need to take place in the main thread.

This patch adds a versatile function to do those actions in a probably best
possible way.