Fix the liveinst install/uninstall hooks
Use the autoconf utility macros in the install hook.

Remove the uninstall hook entirely, because 1) removing all the parent
directories of $(DESTDIR)$(bindir) looks like the kind of thing we
really shouldn't try to do, and 2) there was a syntax error in it
anyway. Automake uninstall targets don't remove directories on the
assumption that if someone really wants $bindir gone they can do it
themselves, so let's follow that and ignore the problem.

Related: rhbz#1121678