Fix the handling of files generated for xgettext
Include the generated files in the archive. Writing files to the source
directory when building from an unpacked distribution archive is a
problem, since VPATH builds expect everything to be written to
$builddir. Gettext insists that all POTFILES be under $srcdir, so
generate the files we need during make dist and pack them up.

To make these dist rules easier, use absolute paths everywhere in
po/Rules-extract. This removes the restriction of target filenames
needing to start with $top_srcdir relative to po. File names can be any
path, either absolute or relative to po, and make will figure out the
rest.