Make _yum.preconf setup atomic (#1028245)
All _yum.preconf operations should be inside of _resetYum, inside a
_yum_lock. In order to do this we need to pass the releasever to the
reset method. I broke up _configureBaseRepo so that it gets the
releasever first, resets the yum object, and then tries to setup the
base repo. The second call to _setupInstallDevice is needed because the
payload.reset tears down the mount. You can't skip doing this without
causing problems switching between an ISO and a http source.

This should eliminate the possibility of other threads calling _yum.conf
while _yum.preconf still needs releasever set.

It is *very* important that _yum.preconf not be touched outside of the
_resetYum method's _yum_lock block.