A nice decorator making Anaconda's GUI more responsive
Related: rhbz#1065716

In many cases we don't want to run expensive callbacks every time a signal is
emitted. We can just schedule the callback to be run later and let the changes
accumulate in the meantime.

For example in the AddLayout dialog, we don't want to refilter the list after
every single change of the entry. Doing so freezes GUI for quite a long time
leaving user with what seems to be a hanged installer.

This way we can schedule actions in a nice way, keeping quite short delay
between changes and callbacks and ensuring callbacks are run with some given
upper bound. Also, by letting the main loop run before the callback, we may set
the busy cursor to indicate something is being processed.

Port of commit 3e8b073ce8859b69e8b53d35a19615c3323fa500 from master.