Change source spoke proxy handling to use local copy (#967805)
The ProxyDialog was using self.data.method.proxy, but if the method has
been changed that will be the wrong one to use.

Instead use a source spoke variable to keep track of the proxy url, pass
this to the dialog for editing (or creation) and set it on the correct
methods in _method_changed.