Allow having unique thread names with given prefix
We need to have unique thread names, but sometimes we need to run the same
action multiple times. Instead of inventing the whell over and over again in
multiple places, we may build-in the functionality in the AnacondaThread class
itself.

Also return thread name when it is added to the ThreadManager.