Make thread manager operations atomic (#1029898)
The thread manipulation in thread manager needs to
be atomic to get rid of nasty race conditions.
For example, without the lock, it was possible
to get a thread object that was not yet started
(as threads are first added to the tracking dict
and then started) and try to join it,
resulting in an exception.
And it was also possible that wait would fail,
as exists & get were called separately,
introducing yet other race condition.
Also for clarity, rename raise_error -> raise_if_error.