Handle non-leaf btrfs volumes with mountpoints. (#1026210)
btrfs allows mounting of non-leaf devices, which is unusual.

This could be a top-level volume or a subvolume that contains other
subvolumes.

The volume could have a mountpoint in a previous installation or it
could have one assigned via kickstart, so we must be able to handle
showing them.

Since the btrfs volume is treated as a container the only
configuration allowed on one is setting a mountpoint or running the
modify container dialog.

(cherry picked from commit 71b081b7f34105d1a35696a8ea9808fbe1718cb7)