Global screenshot support (#1025038)
Previously, Anaconda was not able to take
a screenshot of foreign windows, like for
example the connection editor window or
the keymap window.
This has been fixed by grabbing the content
of the root window and by using the keybinder
global shortcut library, which registers
shortcuts that work regardless of window focus.
As a result, it is now possible to take
a screenshot of any window at any time.

Related: rhbz#752051