Remove base_repo cache (#1011555)
The base repo needs to either be valid or None. So there is no point in
having a separate valid flag.

This renames the decorator to @refresh_base_repo, it now sets _base_repo
to None while inside the decorated methos, and refreshes it at the end.
This way any calls that check base_repo during this time will get None
and they do not depend on _yum_lock at all.

It also moves the decorator a bit. There were several places where it
was being called while inside another decorated method which doesn't
seem like a good thing to do.

This also returns to using baseRepo as a property.