Round float values coming from the Gtk stack (#1013586)
When we put some float value to a TreeStore or when we ask some widget for
a float value it is C-style float with a lot of digits causing problems in our
arithmetics. Since we probably don't have to care about hundredths of megabytes,
we can safely round those values.