Use level objects instead of level integer codes.
Get min_members through the raid object obtained by mdraid.getRaidLevel
rather than passing the raid number to an mdraid procedure.

Previously, if the raid_level had been "container"
an MDRaidLevel exception would have been thrown. Now, there is no exception,
and the corresponding min_members field is 1. In the current interface,
the raid level is always one of the standard RAID levels, so this can never
happen.

Also, previously the error message would always contain the raid_level value
to describe the raid_level. Now it user the raid level's canonical name instead.
However, that seems to be the same as the name that is constructed in the
existing implementation, so expect no observable difference.

Signed-off-by: mulhern <amulhern@redhat.com>