Update source on errors (#1030997)
If the source spoke has errors, treat it as if the source has been
changed and re-fetch metadata. eg. If the network was down this allows
the users to force it to fetch the metadata by re-entering the spoke
once the network is back up.

Resolves: rhbz#1030997
(cherry picked from commit 1e49b72b682b82aca8603763b316b016af8dc603)