Fetch older valid releases
If a package has not bee updated for a new Fedora release,
if might not have a build for the given release.
But the older build will still be tagged the new release
and will be the valid package for the new Fedora.
Makeupdates needs to be able to handle that.