Store older valid packages in separate folder
If a package from older release is used
(as there is no build for the current release
if the package was not updated for the current release),
store it in a separate folder called for_<release id>.
Otherwise the package would be downloaded again each time,
making the updates image creation slow.
Without the subfolder makeupdates would not be enable to distinguish
which cached packages are outdated and should not be used and which
packages just don't have a new build and should be used.