Change how we test if the GUI is available in the anaconda script.
Simply trying to import early on means we get the "Unable to init server" and
XDG_RUNTIME_DIR errors.  Instead, test to see if the GUI is available to be
imported (which it will if it's included in the tree).  We can still fall back
to text mode if there are X problems, though.

Related: rhbz#1078868