Use abstract base classes for mixins.
This is only a beginning, but hopefully enough to get the idea across.
The purpose of this is twofold: For mixins/interfaces/whatever that do
not make sense unless inherited by another class, don't allow the class
to be instantiated. For mixins/interfaces/whatever that depend on
properties that are implemented in the inheriting class or its other
subclasses, make the relationship explicit through the use of abstract
properties.

SourceSwitchHandler is a good example, and the catalyst for this.
SourceSwitchHandler manipulates self.data, which is provided in the
subclass by way of EditTUISpoke -> UIObject. This kind of a
relationship is flagged by pylint since .data is not provided by
SourceSwitchHandler or any of its subclasses. b5df3a90 attempted to
remedy this by storing a copy of data in SourceSwitchHandler, but that
commit is incorrect: self.data in SourceSwitchHandler will resolve to
UIObject.data when called from a class that inherits from both, and
UIObject.data is a read-only property. This commit instead makes
SourceSwitchHandler an abstract base class. It implements a data
property of its own but requires that the property be overridden by
another class. This way no class can inherit from SourceSwitchHandler
without providing a data property.