Fix the botched helperization of StorageChecker
Apparently I smashed "delete" a couple of times while modifying
checkStorage for use as an abstract base class. Also, StorageChecker
really does need to send messages for StorageSpoke and not whatever
class is implementing it, since CustomPartitioningSpoke is not on a hub.
Added a constructor parameter for the spoke name so StorageChecker isn't
totally GUI specific.

Also also I completely misinterpreted how CustomPartitioningSpoke and
StorageSpoke communicate errors via StorageChecker, so reverted that
part and made errors and warnings class attributes on StorageChecker
again.