Move a bunch of abstract methods from Payload to PackagePayload
A lot of things defined for Payload only make sense for PackagePayloads.
This only moves the methods that raise NotImplementedError: any code
that incorrectly accesses one of these properties for a Payload class
that does not inherit from PackagePayload will just get an
AttributeError now instead of a NotImplementedError.  There's probably a
lot more methods that could be moved to better places, but those will
need a closer look.