Tweak DiskOverview spacing a little bit (#1014671).
With the extra options on the storage spoke, DiskOverview icons need to be
smaller.  However, that then requires shrinking the spacing so the words
don't look so dumb next to the smaller icon.

We then also need to make the checkmark icon larger so it doesn't get scaled
down to some tiny thing on the disk icon, and make its check white instead of
transparent so it shows up better.