Fix and ignore markup warnings where appropriate
resize.py and network.py had a spot each that needed markup escaped.
Ignore the other warning in network.py (the string is markup and should
stay as such) and the warning in cart.py (it's a decimal parameter type
but figuring that out in the pylint module is hard).