Use blivet.size.Size for all size quantities.
In addition to resolving a host of long-standing issues with handling of
size units, this changes the default unit in the GUI from MB
(SI/decimal) to MiB (IEC/binary). The reason for this is that everything
is binary at the bottom of the stack. I do not care what disk
manufacturers write on their boxes. Disk sectors are normally 512 bytes
-- not 500 bytes. Sizes of disk partitions will rarely look reasonable
expressed as decimal-prefixed units since an even number in base-10 is
almost never evenly divisible by 512. Since we need to be able to detect
resize based on the contents of the size entry, regardless of units, we
cannot lose any precision in our display of sizes, which rules out lying
to users to make it seem like their decimal-prefixed sizes can be
attained. Sizes can still be expressed in decimal units, but we
will display binary units.

(cherry picked from commit 41f87df183d174322aead7f0cd2fe98a914f54e9)

Related: rhbz#1075561