Check for ready before baseRepo in completed (#1044985)
The completed spoke method can be called before the spoke has finished
initializing itself in cases where the spoke launches its initialize
method in a thread. This can lead to attributes being accessed before
they are ready (namely yum related items).

This adds a check for self.ready to the source and software completed
methods and re-orders the checks so that ready will short out access to
problem attributes like baseRepo.

It also adds some warning docstrings so we don't forget why this can
be a problem.

Resolves: rhbz#1044985