Remove the UID and GID maximums. (#978846)
UIDs and GIDs can be greater than 32535, which seems like kind of a weird limit
anyway, and libuser doesn't configure or enforce any kind of maximum, so
just remove the upper limits. If a user or group is created with a
UID or GID that doesn't fit in uid_t or gid_t, libuser will fail and
anaconda will log a message.