Split the po-based translation code into a separate file.
This adds a directory for common test code that is added to PYTHONPATH
by testenv.sh.

Changed the handling of plural strings in PODict so that all cases of
the string are returned.  Added a check in check_accelerators that the
accelerator for a plural string is the same for all cases of the string.