Do not change sensitivity of label field (#1038590)
Related: rhbz#1038590

This is an unreliable device because the user can:
1) Set a label for a labelable filesystem.
2) Change that filesystem to a non-labelable filesystem
3) Save it all.

To fix it it would be necessary to have an interaction between the filesystem
choice and the label field, deleting the label text when the filesystem is
changed. This would drive me, if I were a user, nuts.

So, instead, just wait until the user presses the accept button, and then tell
them everything that's wrong with their choice of label, including possibly,
its very existence.

Signed-off-by: mulhern <amulhern@redhat.com>