Verify that designated label can be set (#1038590)
Resolves: rhbz#1038590

Once the user has pressed accept verify that the label they have designated,
if it is changed from the previous label, can actually be set.

Display an informative message and reset values if not.

One variable name was changed in the process and had to be substituted in
a couple of places.

Bumped required blivet version.

Signed-off-by: mulhern <amulhern@redhat.com>