handle "ks=cdrom[:<path>]" on systems with multiple CDs (#1049237)
We can't just use /dev/cdrom - what if that's the wrong drive?
Instead, try every cdrom device that appears, until we find the one that
contains the requested file (or /ks.cfg if no path was given).

Also, factor out the bit that checks for "ks=bd:..." and clarify its
error message.