Allow going to the reclaim dialog even for autopart (#1014671).
The previous set of patches had one shortcoming:  If you had enough space to
install and were doing autopart, but still wanted to reclaim space, you could
not get to that dialog.  While I think relatively few people will fall into
this path, we spent a lot of time making it work previously and the reclaim
dialog is the nicest place to shrink other partitions.